Adds the ability to logout
Old session tokens are stored as part of a revocation strategy.
There is a bit of cruft lying around from this PR but in the
intrest of moving fast I'm leaving it.

I'm not happy with the data serialization.  For example, the auth token
is sometimes passed in the header, and sometimes passed in the body (for example
when logging out).  I wasn't able to add the token to the SessionSerializer
for the Netflix Json_FastApi so it's not used for sessions which
is also not ideal.  Whatever.